requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0
numpy>=1.24.0
Pillow>=9.0.0
requests>=2.28.0
gtts>=2.5.0

